+++ fiddles/aurelia/fiddle-0001-SimpleComponent/src/app.js
@@ -1,8 +1,0 @@
-export class App {
-  firstName = 'John';
-  lastName = 'Doe';
-
-  sayHello() {
-    alert(`Hello ${this.firstName} ${this.lastName}. Nice to meet you.`);
-  }
-}