+++ fiddles/aurelia/fiddle-0002-SimpleDependencyInjection/src/app.js
@@ -1,7 +1,0 @@
-import {inject} from 'aurelia-framework';
-import {Dependency} from './dependency';
-
-@inject(Dependency)
-export class App {
-
-}