+++ fiddles/aurelia/fiddle-0002-SimpleDependencyInjection/src/dependency.js
@@ -1,6 +1,0 @@
-import {customElement} from 'aurelia-framework';
-
-@customElement('dependency')
-export class Dependency {
-    message = 'I am a dependency';
-}